% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_traffic.R
\name{aircraft_movements}
\alias{aircraft_movements}
\title{Australian aircraft movements}
\usage{
aircraft_movements()
}
\value{
A tidy data frame of flights passing through Australian airports
}
\description{
Accesses Australian aircraft traffic data from
    (\url{https://bitre.gov.au/publications/ongoing/airport_traffic_data.aspx}).
    Contains the counts of inbound and outbound, domestic and international flights
    from Australian airports, indexed by year.
}
\examples{
\dontrun{
aircraft_movements()
}
}
