% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_traffic.R
\name{airport_passengers}
\alias{airport_passengers}
\title{Australian airport passenger traffic}
\usage{
airport_passengers()
}
\value{
A tidy data frame of passengers passing through Australian airports
}
\description{
Accesses Australian airport passenger traffic data from
    (\url{https://bitre.gov.au/publications/ongoing/airport_traffic_data.aspx}).
    Contains the counts of inbound and outbound, domestic and international revenue
    passengers from Australian airports, indexed by year.
}
\examples{
/dontrun{
airport_passengers()
}
}
