% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports.R
\name{airports}
\alias{airports}
\title{Airport locations}
\usage{
airports()
}
\value{
Returns a dataframe of Australian airports with latitudes and longitude.
}
\description{
This function provides the Name, code and latitude and longitude (in degrees) of
    every airport in Australia or with a flight to Australia. Data is obtained from
    (\url{https://bitre.gov.au/statistics/aviation/}).
}
\examples{
\dontrun{
airports()
}
}
