% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_distances.R
\name{flightdistances}
\alias{flightdistances}
\title{Australian flight distances}
\usage{
flightdistances()
}
\value{
A dataframe of uplift and destination airports (with codes), and the great circle distances
    between them in kilometres (km), nautical miles (nm), and statuteiles (miles).
}
\description{
This function returns a data frame of flight distances to and from each airport in Australia
     (domestic and international). Data is obtained from
    (\url{https://bitre.gov.au/statistics/aviation/}).
}
\examples{
\dontrun{
flightdistances()
}
}
