% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_traffic.R
\name{international_freight}
\alias{international_freight}
\title{Australian international freight data}
\usage{
international_freight()
}
\value{
A dataframe of international freight data through Australian airports
}
\description{
Accesses Australian international freight data from
    (\url{https://bitre.gov.au/publications/ongoing/airport_traffic_data.aspx}).
    Contains inbound and outbound freight and mail in tonnes from Australian
    airports, indexed by year.
}
\examples{
/dontrun{
international_freight()
}
}
