% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozflights-package.R
\docType{package}
\name{ozflights}
\alias{ozflights}
\alias{ozflights-package}
\title{ozflights}
\description{
A package to obtain Australian airport traffic data between 1985-2016 from BITRE.
Contains international freight data and both international and domestic data on number
of passengers, and flight movements - for both regional and metropolitan airports.
Access raw data at \url{https://bitre.gov.au/publications/ongoing/airport_traffic_data.aspx}

